pyproject.toml
[project]
name = "thelastriff"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "mariaeduardanichelle", email = "madunichelle@gmail.com"},
]
dependencies = [
    "django>=5.1.1",
    "djangorestframework>=3.15.2",
    "django-cors-headers>=4.4.0",
    "djangorestframework-simplejwt>=5.3.1",
    "setuptools>=75.1.0",
    "Pillow>=10.4.0",
    "python-magic-bin>=0.4.14",
    "drf-spectacular>=0.27.2",
    "python-dotenv>=1.0.1",
    "python-decouple>=3.8",
    "netifaces>=0.11.0",
    "dj-database-url>=2.2.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.scripts]
pre_dev = "python ./scripts/set_my_ip.py"
dev = "python manage.py runserver 0.0.0.0:19003"